Convert commands from morph to recipe
This includes some basic substitutions ($DESTDIR to ${D}, $PREFIX to
${prefix}), merging commands from the build-system,
and appending those commands to recipes.

There is the possibility of instead taking the build-system and
inheriting an appropriate bbclass, but it won't be general for all our
build-systems (we do odd things with the 'verbatim' build-system),
and I haven't decided whether it's a good idea.